+++ processing/src/test/java/org/apache/druid/segment/IndexMergerV9WithSpatialIndexTest.java
@@ -542,7 +542,7 @@ public class IndexMergerV9WithSpatialIndexTest extends InitializedNullHandlingTe
                                   .filters(
                                       new SpatialDimFilter(
                                           "dim.geo",
-                                          new RadiusBound(new float[]{0.0f, 0.0f}, 5)
+                                          new RadiusBound(new double[]{0.0f, 0.0f}, 5)
                                       )
                                   )
                                   .aggregators(
@@ -594,7 +594,7 @@ public class IndexMergerV9WithSpatialIndexTest extends InitializedNullHandlingTe
                                   .filters(
                                       new SpatialDimFilter(
                                           "spatialIsRad",
-                                          new RadiusBound(new float[]{0.0f, 0.0f}, 5)
+                                          new RadiusBound(new double[]{0.0f, 0.0f}, 5)
                                       )
                                   )
                                   .aggregators(
@@ -645,7 +645,7 @@ public class IndexMergerV9WithSpatialIndexTest extends InitializedNullHandlingTe
                                   .filters(
                                       new SpatialDimFilter(
                                           "dim.geo",
-                                          new RectangularBound(new float[]{0.0f, 0.0f}, new float[]{9.0f, 9.0f})
+                                          new RectangularBound(new double[]{0.0f, 0.0f}, new double[]{9.0f, 9.0f})
                                       )
                                   )
                                   .aggregators(
