+++ processing/src/test/java/org/apache/druid/segment/filter/SpatialFilterTest.java
@@ -551,7 +551,7 @@ public class SpatialFilterTest extends InitializedNullHandlingTest
                                   .filters(
                                       new SpatialDimFilter(
                                           "dim.geo",
-                                          new RadiusBound(new float[]{0.0f, 0.0f}, 5)
+                                          new RadiusBound(new double[]{0.0f, 0.0f}, 5)
                                       )
                                   )
                                   .aggregators(
@@ -603,7 +603,7 @@ public class SpatialFilterTest extends InitializedNullHandlingTest
                                   .filters(
                                       new SpatialDimFilter(
                                           "spatialIsRad",
-                                          new RadiusBound(new float[]{0.0f, 0.0f}, 5)
+                                          new RadiusBound(new double[]{0.0f, 0.0f}, 5)
                                       )
                                   )
                                   .aggregators(
@@ -654,7 +654,7 @@ public class SpatialFilterTest extends InitializedNullHandlingTest
                                   .filters(
                                       new SpatialDimFilter(
                                           "dim.geo",
-                                          new RectangularBound(new float[]{0.0f, 0.0f}, new float[]{9.0f, 9.0f})
+                                          new RectangularBound(new double[]{0.0f, 0.0f}, new double[]{9.0f, 9.0f})
                                       )
                                   )
                                   .aggregators(
